+++ react/features/base/config/index.js
@@ -1,0 +1,4 @@
+export * from './actions';
+export * from './actionTypes';
+
+import './reducer';
